\name{formulaExpand}
\alias{formulaExpand}
\title{Function to expand a dot (.) to represent all variables in an XDF (DF)}
\usage{
  formulaExpand(formula, data)
}
\arguments{
  \item{data}{xdf file}

  \item{formula}{???}
}
\description{
  Function to expand a dot (.) to represent all variables
  in an XDF (DF)
}
\seealso{
  Other Compatibility functions:
  \code{\link{range2formula}}
}

