\name{isLogFormula}
\alias{isLogFormula}
\title{Checks if dependent variable is log transformed.}
\usage{
isLogFormula(object)
}
\arguments{
  \item{object}{An object, either a formula, or containing
  a formula that can be accessed with
  \code{formula(object)}}
}
\description{
Checks if dependent variable is log transformed.
}

