\name{range2formula}
\alias{range2formula}
\title{Function to create a formula from a range on an XDF (DF)}
\usage{
  range2formula(range, data)
}
\arguments{
  \item{data}{xdf file}

  \item{range}{Numeric vector}
}
\description{
  Function to create a formula from a range on an XDF (DF)
}
\seealso{
  Other Compatibility functions:
  \code{\link{formulaExpand}}
}

