\docType{package}
\name{RevoEncancements-package}
\alias{RevoEncancements-package}
\alias{RevoEnhancements}
\title{Adds functionality to RevoScaleR}
\description{
  This package adds functionality to RevoScaleR. All
  functions require, and are enhancements to, [Revolution R
  Enterprise](http://www.revolutionanalytics.com/products/revolution-enterprise.php).
}
\details{
  This program is free software; you can redistribute it
  and/or modify it under the terms of the GNU General
  Public License as published by the Free Software
  Foundation; either version 2 or 3 of the License (at your
  option).

  This program is distributed in the hope that it will be
  useful, but WITHOUT ANY WARRANTY; without even the
  implied warranty of MERCHANTABILITY or FITNESS FOR A
  PARTICULAR PURPOSE.  See the GNU General Public License
  for more details.

  A copy of the GNU General Public License is available at
  http://www.r-project.org/Licenses/

  Areas of enhancement are:

  \itemize{ \item{CRAN R Compatability} \item{Big Data
  Graphics} \item{Big Data Mining} \item{Discretization}
  \item{Variable Importance / Selection} \item{Models}
  \item{Scoring} \item{...} }

  Some functions you might find useful:

  Compatibility and utility functions: \itemize{
  \item{\code{\link{formulaExpand}}}
  \item{\code{\link{range2formula}}} }

  Modeling: \itemize{ \item{\code{\link{rxKS}}}
  \item{\code{\link{rxTTest}}} }

  Data visualisation: \itemize{
  \item{\code{\link{rxBoxPlot}}}
  \item{\code{\link{rxHexBin}}} }

  Data mining: \itemize{ \item{\code{\link{rxDiscretize}}}
  \item{\code{\link{rxSample2Df}}}
  \item{\code{\link{rxTreeDiscretize}}}
  \item{\code{\link{rxF1score}}}
  \item{\code{\link{rxLinPredError}}} }
}
\author{
  Derek Norton \email{derek.norton@revolutionanalytics.com}
  and Andrie de Vries
  \email{andrie@revolutionanalytics.com}
}
\keyword{package}

