\name{rxBoxPlot}
\alias{rxBoxPlot}
\title{Function to create a boxplot from an XDF file.}
\usage{
rxBoxPlot(formula, data, rowSelection = NULL,
  blocksPerRead = rxGetOption("blocksPerRead"), title = NULL,
  subtitle = NULL, xTitle = NULL, yTitle = NULL,
  boxFillColor = "skyblue", boxOutlineColor = "black",
  outlierColor = NULL, outlierStyle = 1, outlierSize = 1,
  outlierBackground = NULL, plotAreaColor = "gray90", gridColor = "white",
  gridLineWidth = 1, gridLineStyle = "solid", whiskerLineType = "solid",
  whiskerLineWidth = 1, whiskerLineColor = "black",
  medianLineType = "solid", medianLineWidth = 1,
  medianLineColor = "black", reportProgress = rxGetOption("reportProgress"))
}
\arguments{
  \item{formula}{???}

  \item{data}{xdf file}

  \item{rowSelection}{???}

  \item{blocksPerRead}{???}

  \item{title}{???}

  \item{subtitle}{???}

  \item{xTitle}{???}

  \item{yTitle}{???}

  \item{boxFillColor}{???}

  \item{boxOutlineColor}{???}

  \item{outlierColor}{???}

  \item{outlierStyle}{???}

  \item{outlierSize}{???}

  \item{outlierBackground}{???}

  \item{plotAreaColor}{???}

  \item{gridColor}{???}

  \item{gridLineWidth}{???}

  \item{gridLineStyle}{???}

  \item{whiskerLineType}{???}

  \item{whiskerLineWidth}{???}

  \item{whiskerLineColor}{???}

  \item{medianLineType}{???}

  \item{medianLineWidth}{???}

  \item{medianLineColor}{???}

  \item{reportProgress}{???}
}
\description{
Function to create a boxplot from an XDF file.
}

