% Generated by roxygen2 (4.0.0): do not edit by hand
\name{rxKS.rxLorenz}
\alias{rxKS.rxLorenz}
\title{Function to calculate the KS Statistic.}
\usage{
\method{rxKS}{rxLorenz}(x)
}
\arguments{
\item{x}{???}
}
\description{
Function to calculate the KS Statistic.
}
\seealso{
Other Modelling functions: \code{\link{rxTTest}}
}

