% Generated by roxygen2 (4.0.0): do not edit by hand
\name{woe}
\alias{woe}
\title{Calculates weight of evidence.}
\usage{
woe(dependent, group, iv = FALSE, breaks = 10, probs = seq(0, 1, by =
  1/breaks), na.rm = TRUE)
}
\arguments{
\item{dependent}{logical or numeric vector in range [0, 1], i.e. a classifier.}

\item{group}{independent variable. Numeric vector of same length as \code{dependent}}
}
\description{
Calculates weight of evidence, similar to how SAS does it.
}
\details{
Information value is defined as IVx =∑_1:10 (bad_i − good_i) log (bad_i * good_i)
}
\references{
http://opinions5.blogspot.co.uk/2011/09/information-value.html

http://support.sas.com/resources/papers/proceedings13/095-2013.pdf
}

