\name{formulaExpand}
\alias{formulaExpand}
\title{Expand a . to represent all variables in an XDF (DF).}
\usage{
formulaExpand(formula, data)
}
\arguments{
  \item{formula}{formula to be expanded}

  \item{data}{XDF file or data.frame from which to create
  new formula}
}
\description{
This function allows the use of . (dot) expansion in functions that would
otherwise not allow it.
}
\seealso{
Other Compatibility functions: \code{\link{range2formula}}
}
\keyword{expand,}
\keyword{formula}

