\name{range2formula}
\alias{range2formula}
\title{Create a formula from a range on an XDF (DF).}
\usage{
range2formula(range, data)
}
\arguments{
  \item{range}{numeric range of values to use in formula}

  \item{data}{XDF file or data.frame from which to create
  new formula}
}
\description{
This function allows the specification of a range of variable instead of a formula
}
\seealso{
Other Compatibility functions: \code{\link{formulaExpand}}
}
\keyword{formula}
\keyword{range,}

